Remove spurious "@" from install rule
The makefile.am install rule tried to output a banner
on installation. In order to avoid echoing the "echo"
command, it was prefixed with "@", but this was done
for _every_ line. The trailing escape ("\") meant that
as far as Make was concerned, it was a single line with
lots of "@" symbols, all of which were passed to the
shell, which then errored out.

It's harrder to explain than to fix!